requirements.txt
yfinance>=0.2.55
pandas>=2.2.0
numpy>=2.0.0
ta>=0.11.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
plotly>=5.24.1
dash>=2.18.2
dash-bootstrap-components>=1.6.0
APScheduler==3.10.4
python-dotenv==1.0.1
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
